Modifying color scheme
