+++ src/components/Avatar.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import MUIAvatar from 'material-ui/Avatar';
+import getColor from '../utils/color-form';
+import titleInitials from '../utils/title-initials';
+
+const Avatar = ({ colorFrom, children, ...rest }) => (
+  <MUIAvatar style={{ backgroundColor: getColor(colorFrom)}} {...rest}>
+    {titleInitials(children)}
+  </MUIAvatar>
+);
+
+
+export default Avatar;
