+++ src/utils/title-initials.js
@@ -4,7 +4,6 @@ export default function titleInitials(title) {
       .split(' ', 2)
       .map(word => word[0])
       .map(chat => chat.toUpperCase())
-      .slice(0, 2)
       .join('');
   } catch(e) {
     console.error(e);
